refactor: move CLI logic into `jazz-tools/dev`
